The work continues

+++ app/src/main/java/com/example/suai_pocket_light/SUAIRaspElement.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.*
 @Serializable
 data class SUAIRaspElement(
     val ItemId: Int,
-    val Week: Int,
-    val Day: Int,
+    val Week: Byte,
+    val Day: Byte,
     val Less: Int,
     val Build: String,
     val Rooms: String,
